feat(get_pydantic): Add support for (once) nested fk_as_int
This way "fk_as_int={"item__category"} should generate a nested item model in which the related category is not a nested model but an ID/PositiveInt. I tried to allow for infinite nesting in fk_as_int but didn't manage to implement it, in fact idk if it's even possible to do with the current relation algorithm.