Support python 3.12 (#1395)
* typing: remove types-pkg-resources package

The package has been yanked on pypi as it has been superceded by
types-setuptools. However, setuptools now provides type-hints for the
pkg-resources package, so it is no longer needed. Additionally, it does
not appear to be used in the project at all, so it appears vestigial.

* fix: recursive_guard is a KW_ONLY arg in 3.12.4+

Python 3.9 added recursive_guard to ForwardRef._evaluate to prevent
infinite recursion of recursive types. This is a private method, and no
API contract is provided. So, when the API changed in the 3.12.4 release
they did not bother to document it, or highlight that it is a breaking
release.

* deps: bump setuptools to ^75.1.0

Versions of setuptools prior to 70.0.0 were vulnerable to a remote
execution exploit documented in CVE-2024-6345-setuptools.

* ci: Add python-3.12 to the test matrix.

* ci: Bump CI poetry version.

---------

Co-authored-by: collerek <collerek@gmail.com>